update product route done

+++ app/forms/product_form.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField
+from wtforms import StringField, IntegerField, FloatField
 from wtforms.validators import DataRequired, ValidationError, Length, URL
 from app.models import Product
 
@@ -7,4 +7,4 @@ class ProductForm(FlaskForm):
     title = StringField('Title', validators=[DataRequired(),Length(min=4, max=150) ])
     description = StringField('Description', validators =[Length(max=1000)])
     image = StringField('Image', validators=[DataRequired(), URL()])
-    price = IntegerField('Price', validators=[DataRequired()])
+    price = FloatField('Price', validators=[DataRequired()])
